Added unit tests for state-editor and answer-group-editor (#13334)
* Added specs

* Added tests

* Lint

* Lint

* Fixed tests

* removed fdescribe

* Removed file name from NOT_FULLY_COVERED_FILENAMES

* Fix e2e

* Addressed review
