--- core/templates/components/state-editor/state-editor.directive.html
@@ -1,112 +1,0 @@
-<md-card class="oppia-editor-card-with-avatar">
-  <div class="state-content-header-container" ng-click="toggleConceptCard()">
-    <div class="state-content-header oppia-mobile-collapsible-card-header">
-      <h3 class="oppia-exp-content-card-header">Content</h3>
-      <i class="fa fa-caret-down"
-         ng-if="!conceptCardIsShown"
-         aria-hidden="true">
-      </i>
-      <i class="fa fa-caret-up"
-         ng-if="conceptCardIsShown"
-         aria-hidden="true">
-      </i>
-    </div>
-  </div>
-  <div class="oppia-mobile-collapsible-card-content">
-    <div class="oppia-editor-card-body oppia-editor-concept-container" ng-if="conceptCardIsShown">
-      <img ng-if="!windowIsNarrow" ng-src="<[oppiaBlackImgUrl]>" alt="" class="oppia-editor-card-avatar">
-      <div id="tutorialStateContent" class="oppia-editor-card-section state-content-editor-parent-container">
-        <state-content-editor on-save-state-content="onSaveStateContent"
-                              show-mark-all-audio-as-needing-update-modal-if-required="showMarkAllAudioAsNeedingUpdateModalIfRequired"
-                              state-content-placeholder="getStateContentPlaceholder()"
-                              state-content-save-button-placeholder="getStateContentSaveButtonPlaceholder()">
-        </state-content-editor>
-      </div>
-    </div>
-  </div>
-</md-card>
-
-<div ng-show="isInteractionShown()">
-  <state-interaction-editor on-save-state-content="onSaveStateContent"
-                            on-save-interaction-id="onSaveInteractionId"
-                            on-save-interaction-customization-args="onSaveInteractionCustomizationArgs"
-                            on-save-next-content-id-index="onSaveNextContentIdIndex"
-                            on-save-solution="onSaveSolution"
-                            recompute-graph="recomputeGraph"
-                            show-mark-all-audio-as-needing-update-modal-if-required="showMarkAllAudioAsNeedingUpdateModalIfRequired">
-  </state-interaction-editor>
-</div>
-
-<div ng-hide="!interactionIdIsSet || currentStateIsTerminal">
-  <state-responses add-state="addState"
-                   navigate-to-state="navigateToState"
-                   on-responses-initialized="reinitializeEditor"
-                   on-save-inapplicable-skill-misconception-ids="onSaveInapplicableSkillMisconceptionIds"
-                   on-save-interaction-answer-groups="onSaveInteractionAnswerGroups"
-                   on-save-interaction-default-outcome="onSaveInteractionDefaultOutcome"
-                   on-save-next-content-id-index="onSaveNextContentIdIndex"
-                   on-save-solicit-answer-details="onSaveSolicitAnswerDetails"
-                   refresh-warnings="refreshWarnings"
-                   show-mark-all-audio-as-needing-update-modal-if-required="showMarkAllAudioAsNeedingUpdateModalIfRequired">
-  </state-responses>
-  <div>
-    <state-hints-editor on-save-hints="onSaveHints"
-                        on-save-next-content-id-index="onSaveNextContentIdIndex"
-                        on-save-solution="onSaveSolution"
-                        show-mark-all-audio-as-needing-update-modal-if-required="showMarkAllAudioAsNeedingUpdateModalIfRequired">
-    </state-hints-editor>
-  </div>
-  <div ng-if="currentInteractionCanHaveSolution">
-    <state-solution-editor on-save-solution="onSaveSolution"
-                           refresh-warnings="refreshWarnings"
-                           show-mark-all-audio-as-needing-update-modal-if-required="showMarkAllAudioAsNeedingUpdateModalIfRequired">
-    </state-solution-editor>
-  </div>
-  <div ng-if="explorationIsLinkedToStory">
-    <state-skill-editor (on-save-state-content)="onSaveStateContent($event)"
-                        (on-save-linked-skill-id)="onSaveLinkedSkillId($event)">
-    </state-skill-editor>
-  </div>
-</div>
-<style>
-  state-editor .state-content-header-container {
-    display: none;
-    padding: 0 30px;
-  }
-  state-editor .state-content-header {
-    align-content: center;
-    align-items: center;
-    display: flex;
-    justify-content: space-between;
-    padding: 30px 0 15px;
-  }
-  state-editor .state-content-header i {
-    display: none;
-  }
-  state-editor .oppia-exp-content-card-header {
-    font-size: 18px;
-    margin: 0;
-  }
-  state-editor .state-solution-content-container {
-    padding: 15px 30px;
-  }
-  state-editor .oppia-add-hint-button-container {
-    margin: 5px 0 15px;
-  }
-  state-editor .state-content-editor-parent-container {
-    padding-bottom: 36px;
-    padding-top: 36px;
-  }
-  @media screen and (max-width: 768px) {
-    state-editor .state-content-header-container {
-      display: block;
-      padding: 0;
-    }
-    state-editor .state-content-header {
-      padding: 18px 15px;
-    }
-    state-editor .state-content-header i {
-      display: block;
-    }
-  }
-</style>
