+++ core/templates/components/state-editor/state-responses-editor/state-responses.component.ts
@@ -22,10 +22,10 @@ require(
   'confirm-or-cancel-modal.controller.ts');
 require(
   'components/state-directives/answer-group-editor/' +
-  'answer-group-editor.directive.ts');
+  'answer-group-editor.component.ts');
 require(
   'components/state-directives/response-header/response-header.directive.ts');
-require('components/state-editor/state-editor.directive.ts');
+require('components/state-editor/state-editor.component.ts');
 require(
   'components/state-directives/outcome-editor/' +
   'outcome-destination-editor.directive.ts');
