+++ scripts/check_frontend_test_coverage.py
@@ -40,7 +40,6 @@ EXCLUDED_DIRECTORIES = [
 NOT_FULLY_COVERED_FILENAMES = [
     'angular-html-bind.directive.ts',
     'answer-classification.service.ts',
-    'answer-group-editor.directive.ts',
     'App.ts',
     'audio-player.service.ts',
     'audio-preloader.service.ts',
@@ -216,7 +215,6 @@ NOT_FULLY_COVERED_FILENAMES = [
     'solution-explanation-editor.directive.ts',
     'state-card.model.ts',
     'state-content-editor.directive.ts',
-    'state-editor.directive.ts',
     'state-hints-editor.directive.ts',
     'state-interaction-editor.directive.ts',
     'state-property.service.ts',
